+++ packages/sdk/src/sdk-blockchains/flow/test/sell-item.ts
@@ -1,5 +1,5 @@
 import type { ItemId, OrderId } from "@rarible/types"
-import { toBigNumber } from "@rarible/types"
+import { toBigNumber, toCurrencyId } from "@rarible/types"
 import { FLOW_TESTNET_ACCOUNT_2 } from "@rarible/flow-test-common"
 import type { FlowSell } from "../sell"
 import { convertFlowUnionAddress } from "../common/converters"
@@ -16,6 +16,21 @@ export async function sellItem(sell: FlowSell, itemId: ItemId, priceDecimals: st
 			"@type": "FLOW_FT",
 			contract: testFlowToken,
 		},
+		itemId,
+		originFees: [{ account: convertFlowUnionAddress(FLOW_TESTNET_ACCOUNT_2.address), value: 200 }],
+	})
+	expect(orderId).toBeTruthy()
+	return orderId
+}
+
+export async function sellItemWithCurrencyId(sell: FlowSell, itemId: ItemId, priceDecimals: string): Promise<OrderId> {
+	const { submit } = await sell.sell({
+		collectionId: testFlowCollection,
+	})
+	const orderId = await submit({
+		amount: 1,
+		price: toBigNumber(priceDecimals),
+		currency: toCurrencyId(testFlowToken),
 		itemId,
 		originFees: [{ account: convertFlowUnionAddress(FLOW_TESTNET_ACCOUNT_2.address), value: 200 }],
 	})