+++ packages/sdk/src/sdk-blockchains/tezos/sale.test.ts
@@ -1,9 +1,11 @@
-import { toBigNumber, toContractAddress, toItemId, toUnionAddress } from "@rarible/types"
+import { toBigNumber, toContractAddress, toCurrencyId, toItemId, toUnionAddress } from "@rarible/types"
 import BigNumber from "bignumber.js"
+import type { TezosXTZAssetType } from "@rarible/api-client"
 import { createRaribleSdk } from "../../index"
 import { LogsLevel } from "../../domain"
 import { createTestWallet } from "./test/test-wallet"
 import { awaitForOwnership } from "./test/await-for-ownership"
+import { awaitForOrder } from "./test/await-for-order"
 
 describe("test tezos mint and sell", () => {
 	const sellerWallet = createTestWallet(
@@ -37,6 +39,39 @@ describe("test tezos mint and sell", () => {
 			supply: 1,
 			lazyMint: false,
 		})
+
+		const fillResponse = await buyerSdk.order.buy({ orderId: mintResult.orderId })
+
+		const fillResult = await fillResponse.submit({
+			amount: 1,
+			infiniteApproval: true,
+		})
+		await fillResult.wait()
+
+		const ownership = await awaitForOwnership(
+			buyerSdk,
+			toItemId(mintResult.itemId),
+			await buyerWallet.provider.address()
+		)
+		expect(ownership.value).toBe("1")
+	})
+
+	test("sale NFT with XTZ and with CurrencyId", async () => {
+		const mintAndSellAction = await sellerSdk.nft.mintAndSell({
+			collectionId: toContractAddress(`TEZOS:${nftContract}`),
+		})
+
+		const mintResult = await mintAndSellAction.submit({
+			price: new BigNumber("0.0001"),
+			currency: toCurrencyId("TEZOS:tz1Ke2h7sDdakHJQh8WX4Z372du1KChsksyU"),
+			uri: "ipfs://bafkreiaz7n5zj2qvtwmqnahz7rwt5h37ywqu7znruiyhwuav3rbbxzert4",
+			supply: 1,
+			lazyMint: false,
+		})
+
+		const order = await awaitForOrder(sellerSdk, mintResult.orderId)
+		const takeAssetType = order.take.type as TezosXTZAssetType
+		expect(takeAssetType["@type"]).toEqual("XTZ")
 
 		const fillResponse = await buyerSdk.order.buy({ orderId: mintResult.orderId })