+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.47](https://github.com/rarible/sdk/compare/v0.8.46...v0.8.47) (2022-03-28)
+
+
+### Bug Fixes
+
+* add npm token ([387865d](https://github.com/rarible/sdk/commit/387865d413855e55a37aa6768fb87c521a2462f6))
+* remove lazy mint on polygon ([#169](https://github.com/rarible/sdk/issues/169)) ([e74bc31](https://github.com/rarible/sdk/commit/e74bc314249af7a208c9d4cbae0b2d1e1c752821))
+* tezos sign function and improve deps ([#158](https://github.com/rarible/sdk/issues/158)) ([791e8c8](https://github.com/rarible/sdk/commit/791e8c8cf2cbac84b77c408993134eae79280049))
+* update jenkins file ([56d7641](https://github.com/rarible/sdk/commit/56d76416b99ea5a9f5d4e920e70ce1ab95ae426b))
+
+
+### Features
+
+* fix lint ([c2fd61a](https://github.com/rarible/sdk/commit/c2fd61a2d550202f1fed2c3c58bc3d7eb99bf6cc))
+
+
+
+
+
 ## [0.8.46-hotfix.1](https://github.com/rarible/sdk/compare/v0.8.46-hotfix.0...v0.8.46-hotfix.1) (2022-03-21)